+++ src/net/dash/texttracks.ts
@@ -198,11 +198,14 @@ function TextTrackParser({
   }
 
   if (isInit) {
-    segmentInfos = { time: -1, duration: 0 };
     if (isMP4) {
       const timescale = getMDHDTimescale(responseData as Uint8Array);
       if (timescale > 0) {
-        segmentInfos.timescale = timescale;
+        segmentInfos = {
+          time: -1,
+          duration: 0,
+          timescale,
+        };
       }
     }
     segmentData = undefined;
@@ -272,7 +275,7 @@ function TextTrackParser({
   }
 
   if (nextSegments) {
-    addNextSegments(representation, segmentInfos, nextSegments);
+    addNextSegments(representation, nextSegments, segmentInfos);
   }
   return Observable.of({ segmentData, segmentInfos });
 }